Arparn monster animation

See merge request ilboit/deepdarkdungeon2020!37

+++ core/src/ee/taltech/deepdarkdungeon/Models/BadCharacter.java
@@ -8,14 +8,14 @@ public class BadCharacter extends GameObject {
     String name;
     int health;
     int power;
-    double x;
-    double y;
+    int x;
+    int y;
     double width;
     double height;
     CharacterType characterType;
     Texture picture;
 
-    public BadCharacter(Texture texture, String name, CharacterType characterType, int power, int health, double x, double y, double width, double height) {
+    public BadCharacter(Texture texture, String name, CharacterType characterType, int power, int health, int x, int y, double width, double height) {
         this.x = x;
         this.y = y;
         this.width = width;
