+++ core/src/ee/taltech/deepdarkdungeon/Models/characterClasses/Warrior.java
@@ -5,7 +5,7 @@ import ee.taltech.deepdarkdungeon.Models.GameObject;
 
 public class Warrior extends GameObject {
 
-    public Warrior(Texture picture, String name, int health, int power, double x, double y, double width, double height, CharacterClass characterClass, CharacterType characterType) {
+    public Warrior(Texture picture, String name, int health, int power, int x, int y, double width, double height, CharacterClass characterClass, CharacterType characterType) {
         this.texture = picture;
         this.name = name;
         this.health = health;
